Add files via upload
These are the various images used in the Game. There is a folder inside it with the name 
Characters that include the sprite sheet of the characters used inside the game.

Also there is the images of the various animated buttons that are used with some graphics 
and animations.

These also include the background images of various levels.
You can customize the background but you should specify the width and height in the program
and the level for the proper functioning of the game.